Style for the day block
Added style for the day block support.

Add this to the calendar init

eventSource: function (date_from, date_to, callback) {

}

Return an array of with all days (Objects) between 'date_from' to
'date_to' with the position in the array as the day of the month without
a leading zero.

Format for the day Object

{style: <class_css_name>}